requirement.txt
djangorestframework
django
djoser
mysqlclient